# add VALAST_PROFILE to emit performance profiling information